+++ src/camera.hpp
@@ -20,18 +20,18 @@ public:
 
   Camera(
     CameraType new_type,
-    real64 window_width,
-    real64 window_height
+    uint32 window_width,
+    uint32 window_height
   );
   void move_front_back(real32 sign, real64 dt);
   void move_left_right(real32 sign, real64 dt);
   void move_up_down(real32 sign, real64 dt);
   void update_mouse(glm::vec2 mouse_offset);
-  void update_matrices_perspective(real64 window_width, real64 window_height);
-  void update_matrices_ortho(real64 window_width, real64 window_height);
-  void update_matrices(real64 window_width, real64 window_height);
+  void update_matrices_perspective(uint32 window_width, uint32 window_height);
+  void update_matrices_ortho(uint32 window_width, uint32 window_height);
+  void update_matrices(uint32 window_width, uint32 window_height);
   void update_ui_matrices(
-    real64 window_width, real64 window_height
+    uint32 window_width, uint32 window_height
   );
   static void create_shadow_transforms(
     glm::mat4 shadow_transforms[6 * MAX_N_LIGHTS],
